+++ libsrc2/minc2_private.h
@@ -73,7 +73,7 @@ struct midimension {
   char *name;                   /* Dimension name */
   double *offsets;              /* Offsets (if irregular) */
   double step;                  /* Step size */
-  unsigned long length;         /* Length */
+  unsigned int length;         /* Length */
   double start;                 /* Start value */
   char *units;                  /* Units string */
   double width;                 /* Sample width (if regular) */
